validation

+++ simple-routes/controllers/admin.js
@@ -24,7 +24,7 @@ exports.postAddProduct = (req, res, next) => {
     .save()
     .then(result => {
       // console.log(result);
-      // console.log('Created Product');
+      console.log('Created Product');
       res.redirect('/admin/products');
     })
     .catch(err => {
@@ -62,30 +62,27 @@ exports.postEditProduct = (req, res, next) => {
 
   Product.findById(prodId)
     .then(product => {
-      if(product.userId.toString() !== req.user._id.toString()){
+      if (product.userId.toString() !== req.user._id.toString()) {
         return res.redirect('/');
-      }else{
-        product.title = updatedTitle;
-        product.price = updatedPrice;
-        product.description = updatedDesc;
-        product.imageUrl = updatedImageUrl;
-        return product.save();
       }
-    })
-    .then(result => {
-      // console.log('UPDATED PRODUCT!');
-      res.redirect('/admin/products');
+      product.title = updatedTitle;
+      product.price = updatedPrice;
+      product.description = updatedDesc;
+      product.imageUrl = updatedImageUrl;
+      return product.save().then(result => {
+        console.log('UPDATED PRODUCT!');
+        res.redirect('/admin/products');
+      });
     })
     .catch(err => console.log(err));
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({
-    userId: req.user._id
-  })
+  Product.find({ userId: req.user._id })
     // .select('title price -_id')
     // .populate('userId', 'name')
     .then(products => {
+      console.log(products);
       res.render('admin/products', {
         prods: products,
         pageTitle: 'Admin Products',
@@ -97,13 +94,10 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({
-    _id: prodId,
-    userId: req.user._id
-  })
-  .then(() => {
-    console.log('DESTROYED PRODUCT');
-    res.redirect('/admin/products');
-  })
-  .catch(err => console.log(err));
+  Product.deleteOne({ _id: prodId, userId: req.user._id })
+    .then(() => {
+      console.log('DESTROYED PRODUCT');
+      res.redirect('/admin/products');
+    })
+    .catch(err => console.log(err));
 };
